Major changes related to templates Some major changes to the template language:  Added an absolute ton of functions based on ones in Haskell's Data.List, for processing lists and strings (two copies of most of them, one for lists and one for strings, since a template string is not a list).
Pursuant to this, removed a lot of stuff from the template language's Types.hs:
The Eq and Show instances, and the type-class Bindable.  The removal of the
last makes code elsewhere more complicated, but the code that it complicates
is due to be ported to the template language and migrated into the database
soon, so it doesn't matter.

Also removed getInterfaceStateMVar, replacing it with two separate MVars, one
for template-language bindings and one for CAPTCHAs (the only two modules to
actually use the now-removed generic-interface-state facility).  This meant also
removing the initState functions from each interface, which is fine; I'm trying
to remove the entire interface, after all, so as to avoid the use of plugins
altogether.

I'm checking this in even though it's in a non-compiling state because I'm about
to start something even more radical, converting everything back to one package,
and I want the flexibility to rewind to this point.