Yet another broken commit, this one prior to rewriting the Custard language to use typechecking-in-advance instead of on-demand. Removed the concept of template-expression queries.  All queries are once again pure SQL.
Moved template parameters out of the bindings mvar and into the context.  Since
bindings have been broken for concurrent requests all along, it would be nice
if I could move them in there, too.  But in the short term, template parameters
HAVE to move because they can't be a template list because template lists are
to be homogeneous.

While doing that I went ahead and also moved bindings into the context.

And I renamed what was formerly known as the template language to "Custard",
because it was getting confusing that it was used for things besides templates.

And there are some semantics changes to the language itself.  The lookup()
special form is removed, because to be useful it would need first-class symbols,
which if I want at all I don't want to do as part of this massive spree.  The
query() special form is changed to not return a value.